refactor(core): move checkout rebase mutation
